# Cuda refactor, multi GPU support