This PR is quite large. Its primary goal is to lay the groundwork for the implementation of further CUDA kernels for ggml operations. I am also adding multi GPU support because it's easier to integrate now than it would be at a later point. 

### For Users

Build instructions (Linux):

```
git clone https://github.com/JohannesGaessler/llama.cpp llama.cpp-johannesgaessler
cd llama.cpp-johannesgaessler                               
git fetch
git switch cuda-refactor-8
make LLAMA_CUBLAS=1
```

When compiling with `LLAMA_CUBLAS=1` the program automatically detects the available NVIDIA devices and splits weights proportional to VRAM. There is not yet a CLI argument for setting the tensor splits. The performance increase on my test systems is relatively low (+70% t/s when going from 1x GTX TITAN X to 4x GTX TITAN X). It's possible that there is still a bug that hampers performance. Please do tell me how well (if at all) it works for you. In any case, this PR should already allow you to pool the VRAM of multiple GPUs to load larger models.

### For Developers

~~**This PR is still very much WIP. I will do a refactor to remove artifacts from bad/obsolete design decisions. You can already review the code if you want but many of the flaws are still subject to change.**~~ Should be good now.

On master there are separate functions for invoking CUDA kernels. Apart from invoking the actual CUDA kernels they do other things such as copying data between host and device. This PR adds a template `ggml_cuda_op` that manages

1. the transfer of data between host and device,
2. the dequantization of `src0` (needed for cuBLAS matrix multiplication),
3. the broadcasting of `src1` across `src0` (needed for multiplication),
4. and multi GPU things.

The actual operations now only need to define how the data should be manipulated.

This PR also moves the entry point for invoking CUDA kernels from the ggml function such as `ggml_compute_forward_mul_mat_q_f32` and instead adds a function `ggml_cuda_compute_forward` that is called from `ggml_compute_forward`. For this to work I moved `ggml_task_type` and `ggml_compute_params` from `ggml.c` to `ggml.h`.

This PR adds an int for the layer, an int for the device id, and dedicated device data pointers to `ggml_tensor`. I need these for bookkeeping. I also changed the backends from `GGML_BACKEND_CUDA` and `GGML_BACKEND_OPENCL` to `GGML_BACKEND_GPU` (tensor data on 1 GPU) and `GGML_BACKEND_GPU_SPLIT` (tensor data split across all GPUs). Since I think that we don't want to support the simultaneous use of CUDA and OpenCL it's simpler to just use the same backend types for both implementations and to differentiate via defines.