@JohannesGaessler 2023-05-27
Performance numbers from my test machine with an i5-4570S, 16 GB of RAM @ 1600 MHz, and a GTX 1070 + a GTX 1050 ti:

| Model    | GPU                    |     ms/t |      t/s |
|----------|------------------------|----------|----------|
| 7b q4_0  | GTX 1070               |    71.37 |    14.01 |
| 7b q4_0  | GTX 1070 + GTX 1050 ti |    68.66 |    14.56 |
| 7b q4_0  | GTX 1070               |    71.37 |    14.01 |
| 13b q4_0 | GTX 1070               |   134.19 |     7.45 |
| 13b q4_0 | GTX 1070 + GTX 1050 ti |   128.13 |     7.80 |
| 33b q4_0 | GTX 1070               | Unusable | Unusable |
| 33b q4_0 | GTX 1070 + GTX 1050 ti |   575.12 |     1.74 |

Numbers for single GPU are obtained using the master branch.

Note: previously I was able to run 33b q4_0 with just the GTX 1070 on master; there may be something on master that has increased RAM usage since.