@SlyEcho 2023-05-27
> 33b

You mean 30B? I can run 30B Q4_0 with my 8 GB card with 20 layers loaded only.