@slaren 2023-05-27
I think `ggml_cuda_mul` and `ggml_cuda_mul_mat` can be removed from `ggml-cuda.h` now and made static.