@JohannesGaessler 2023-05-27
I added a comment to explain the weird device to host memcpy for split tensors. Since I as the person to wrote the code won't know: are there other parts of the code that are unintuitive or difficult to understand?