@KerfuffleV2 2023-06-05
```
ggml_init_cublas: found 1 CUDA devices:
  1. NVIDIA GeForce GTX 1060 6GB
```

Works for me on Linux, performance seems the same as master. Identical results with the same seed also. Perplexity (and I'd assume just cuBLAS prompt evaluation in general) is noticeably slower though. (Output abbreviated below.)

### `master`

```
$ taskset -c 0-5 ./perplexity -f /path/wiki.test.raw -m /path/llama-7b.ggmlv3.q4_0.bin -t 6 -ngl 0
system_info: n_threads = 6 / 12 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 1 | VSX = 0 | 
perplexity: calculating perplexity over 655 chunks, batch_size=512
perplexity: 3.66 seconds per pass - ETA 39 minutes
[1]4.4545,[2]4.9402,[3]5.8278
```

***

```
$ taskset -c 0-5 ./perplexity -f /path/wiki.test.raw -m /path/llama-7b.ggmlv3.q4_0.bin -t 6 -ngl 33
system_info: n_threads = 6 / 12 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 1 | VSX = 0 | 
perplexity: calculating perplexity over 655 chunks, batch_size=512
perplexity: 3.37 seconds per pass - ETA 36 minutes
[1]4.4547,[2]4.9404
```


### this pull

```
$ taskset -c 0-5 ./perplexity -f /path/wiki.test.raw -m /path/llama-7b.ggmlv3.q4_0.bin -t 6 -ngl 0
system_info: n_threads = 6 / 12 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 1 | VSX = 0 | 
perplexity: calculating perplexity over 655 chunks, batch_size=512
perplexity: 5.47 seconds per pass - ETA 59 minutes
[1]4.4545,[2]4.9402
```

***

```
$ taskset -c 0-5 ./perplexity -f /path/wiki.test.raw -m /path/llama-7b.ggmlv3.q4_0.bin -t 6 -ngl 33
system_info: n_threads = 6 / 12 | AVX = 1 | AVX2 = 1 | AVX512 = 0 | AVX512_VBMI = 0 | AVX512_VNNI = 0 | FMA = 1 | NEON = 0 | ARM_FMA = 0 | F16C = 1 | FP16_VA = 0 | WASM_SIMD = 0 | BLAS = 1 | SSE3 = 1 | VSX = 0 | 
perplexity: calculating perplexity over 655 chunks, batch_size=512
perplexity: 5.12 seconds per pass - ETA 55 minutes
[1]4.4545,[2]4.9402
```